requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
APScheduler==3.10.4
requests==2.32.3
stripe==10.11.0
openai==1.43.0
Werkzeug==3.0.3
WTForms==3.1.2